Add new modules
