+++ js/card.js
@@ -1,0 +1,2 @@
+// Модуль работы с карточками обьявлений
+// Пункты ТЗ: 2.
