+++ js/filters.js
@@ -1,0 +1,2 @@
+// Модуль с функциями фильтрации данных, сравнение с похожими обьявлениями
+// Пункты ТЗ: 5.8, 5.9, 5.10, 5.11.
