+++ js/map.js
@@ -1,0 +1,2 @@
+// Модуль работы с картой
+// Пункты ТЗ: 4.
