+++ js/server.js
@@ -1,0 +1,2 @@
+// Модуль работы с сервером, отправка и получение данных
+// Пункты ТЗ: 2.3, 2.4, 5.1, 5.2.
