+++ js/validation.js
@@ -1,0 +1,2 @@
+// Модуль с функциями валидации
+// Пункты ТЗ: 2.5, 3
